Refactor tests for today line
