+++ tests/app.spec.js
@@ -1,19 +1,10 @@
 import { test, expect } from '@playwright/test';
-import fs from 'fs/promises';
-import path from 'path';
-
-const sampleCsvPath = path.join(__dirname, '..', 'csv', 'sample.csv');
+import { loadSampleCsv } from './utils.js';
 
 // テキストエリアにサンプルCSVを入力してレンダリングできるかを確認
 
 test('renders bars from sample CSV', async ({ page }) => {
-  await page.goto('/');
-  // CSV入力モーダルを開いてテキストエリアを表示
-  await page.click('#previewBtn');
-  await page.waitForSelector('#csvInput', { state: 'visible' });
-  const csv = await fs.readFile(sampleCsvPath, 'utf-8');
-  await page.fill('#csvInput', csv);
-  await page.click('#renderBtn');
+  await loadSampleCsv(page);
   // バーが生成されるまで待機し、1つ以上あることを確認
   await page.waitForSelector('#bars .bar');
   const count = await page.locator('#bars .bar').count();
